+++ database/migrations/2020_12_19_091508_create_user_petugas_table.php
@@ -18,7 +18,7 @@ class CreateUserPetugasTable extends Migration
             $table->string("username");
             $table->string("nama");
             $table->string("password");
-            $table->integer("role");
+            $table->bigInteger("role");
             $table->string("email");
             $table->timestamps();
         });
